fix: Add workaround for de-synched checkbox issue
When a user changed the `checked` property (not attribute) of
a `<input type="checkbox">` slotted inside `<turtle-checkbox>` or
`<turtle-toggle-switch>`, the background color changes but icon/knob
does not. This is due to there is no way to listen for DOM property
changes and no events for programmatic property change.

As a workaround, I added a very short transition for the slotted
input and `transitionstart` listener that sync checked state.